Add 'delete own' permission

+++ administrator/components/com_joomgallery/views/images/view.html.php
@@ -89,11 +89,10 @@ class JoomGalleryViewImages extends JoomGalleryView
       JToolbarHelper::divider();
     }
 
-    //if($canDo->get('core.delete'))
-    //{
+    if($canDo->get('core.delete') || $canDo->get('joom.delete.own'))
+    {
       JToolbarHelper::deleteList('', 'remove');
-    //}
-
+    }
   }
 
   /**
